refactor: move sequelize repository tests
